requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.1
filelock==3.9.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
orjson==3.9.7
Pillow==9.3.0
pip==23.2.1
pydantic==2.3.0
pydantic_core==2.6.3
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.28.1
setuptools==68.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.11.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.8.0
ujson==5.8.0
urllib3==1.26.13
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
wheel==0.38.4
